+++ node/resolvers/catalog/category.ts
@@ -22,9 +22,8 @@ export const pathToCategoryHref = (path: string) => {
 interface SafeCategory
   extends Pick<
   Category,
-  'id' | 'name' | 'hasChildren' | 'MetaTagDescription' | 'Title'
+  'id' | 'name' | 'hasChildren' | 'MetaTagDescription' | 'Title' | 'url'
   > {
-  url: string | null
   children: Category[] | null
 }
 
@@ -32,35 +31,15 @@ export const resolvers = {
   Category: {
     cacheId: prop('id'),
 
-    href: async (
-      { id, url }: SafeCategory,
-      _: any,
-      { clients: { catalog } }: Context
-    ) => {
-      if (url == null) {
-        const category = await getCategoryInfo(catalog, id, 4)
-        url = category.url
-      }
-      const path = cleanUrl(url)
-
-      // If the path is `/clothing`, we know that's a department
-      // But if it is `/clothing/shirts`, it's not.
-      return pathToCategoryHref(path)
+    href: async ({ url }: SafeCategory) => {
+      return cleanUrl(url)
     },
 
     metaTagDescription: prop('MetaTagDescription'),
 
     titleTag: prop('Title'),
 
-    slug: async (
-      { id, url }: SafeCategory,
-      _: any,
-      { clients: { catalog } }: Context
-    ) => {
-      if (url == null) {
-        const category = await getCategoryInfo(catalog, id, 4)
-        url = category.url
-      }
+    slug: async ({ url }: SafeCategory) => {
       return url ? lastSegment(url) : null
     },
 
@@ -70,7 +49,7 @@ export const resolvers = {
       { clients: { catalog } }: Context
     ) => {
       if (children == null) {
-        const category = await getCategoryInfo(catalog, id, 4)
+        const category = await getCategoryInfo(catalog, id, 5)
         children = category.children
       }
       return children